Restructure DTOs folder
