--- src/main/java/com/matchbook/sdk/clients/rest/UserReource.java
@@ -1,12 +1,0 @@
-package com.matchbook.sdk.clients.rest;
-
-import com.matchbook.sdk.clients.rest.dtos.LoginRequest;
-import com.matchbook.sdk.clients.rest.dtos.LoginResponse;
-import com.matchbook.sdk.clients.rest.dtos.LogoutResponse;
-
-public interface UserReource {
-
-    LoginResponse login(LoginRequest loginRequest);
-
-    LogoutResponse logout(String sessionToken);
-}
