--- src/main/java/com/matchbook/sdk/clients/rest/dtos/LoginResponse.java
@@ -1,9 +1,0 @@
-package com.matchbook.sdk.clients.rest.dtos;
-
-public class LoginResponse {
-
-    private String sessionToken;
-    private long userId;
-    private Account account;
-
-}
