--- src/main/java/com/matchbook/sdk/clients/rest/dtos/LogoutResponse.java
@@ -1,8 +1,0 @@
-package com.matchbook.sdk.clients.rest.dtos;
-
-public class LogoutResponse {
-
-    private String sessionToken;
-    private Long userId;
-    private String username;
-}
