+++ src/main/java/com/matchbook/sdk/clients/rest/dtos/MatchbookRequest.java
@@ -1,0 +1,5 @@
+package com.matchbook.sdk.clients.rest.dtos;
+
+public interface MatchbookRequest {
+
+}
