+++ src/main/java/com/matchbook/sdk/clients/rest/dtos/user/LoginResponse.java
@@ -1,0 +1,11 @@
+package com.matchbook.sdk.clients.rest.dtos.user;
+
+import com.matchbook.sdk.clients.rest.dtos.MatchbookResponse;
+
+public class LoginResponse implements MatchbookResponse {
+
+    private String sessionToken;
+    private Long userId;
+    private Account account;
+
+}
